refactor: 컨트롤러 제거
